requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
drf-yasg==1.21.10
gunicorn==23.0.0
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
Markdown==3.8.2
packaging==25.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
whitenoise==6.9.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
Pillow==10.4.0
